update degree days controller

+++ app/controllers/application_controller.rb
@@ -71,11 +71,11 @@ class ApplicationController < ActionController::Base
   end
 
   def lat
-    params[:lat]&.to_d&.round(1)
+    params[:lat]&.to_f&.round(1)
   end
 
   def long
-    params[:long]&.to_d&.round(1)
+    params[:long]&.to_f&.round(1)
   end
 
   def lat_range
